Replace double for float

+++ lib/Point.cpp
@@ -11,7 +11,7 @@ Point::Point(): x(0), y(0), z(0){}
  * @param y Value of Y
  * @param z Value of Z
  */
-Point::Point(double x, double y, double z) : x(x), y(y), z(z){}
+Point::Point(float x, float y, float z) : x(x), y(y), z(z){}
 
 /**
  * @brief Transforms Point into String
